+++ src/index.js
@@ -1,4 +1,4 @@
-import osmAuth from 'osm-auth';
+import { osmAuth } from 'osm-auth';
 import defaultOptions from './defaultOptions.json';
 import { getCurrentIsoTimestamp } from 'helpers/time';
 import {
@@ -44,11 +44,16 @@ export default class OsmRequest {
   /**
    * @access public
    * @param {Object} [options] Custom options to apply
-   * @param {string} [options.endpoint] URL of the OSM server to use (defaults to https://www.openstreetmap.org)
-   * @param {boolean} [options.always_authenticated] Set to true if every call to API should use your credentials, false for only requiring calls (defaults to false)
-   * @param {Object} [options.basicauth] Use basic authentication instead of OAuth (not safe)
-   * @param {string} [options.basicauth.user] Username
-   * @param {string} [options.basicauth.pass] Password
+   * @param {string} [options.scope] Scopes separated by a space, see https://wiki.openstreetmap.org/wiki/OAuth#OAuth_2.0 for all scopes (defaults to read_prefs write_prefs write_api write_notes)
+   * @param {string} [options.client_id] The Client ID of your app. To register an app, see https://wiki.openstreetmap.org/wiki/OAuth#OAuth_2.0_2
+   * @param {string} [options.redirect_uri] URL of your app (or url of app code) to rederict to after authentication.
+   * @param {string} [options.access_token] Optional. OAuth2 bearer token to pre-authorize your app.
+   * @param {string} [options.url] URL of the OSM server to use (defaults to https://openstreetmap.org)
+   * @param {string} [options.apiUrl] URL of the OSM API to use (defaults to https://api.openstreetmap.org)
+   * @param {boolean} [options.auto] If true, attempt to authenticate automatically when calling .xhr() or fetch() (default: false)
+   * @param {Object} [options.singlepage] If true, use page redirection instead of a popup (default: false)
+   * @param {string} [options.loading] Function called when auth-related xhr calls start
+   * @param {string} [options.done] Function called when auth-related xhr calls end
    */
   constructor(options = {}) {
     this._options = {
@@ -56,27 +61,29 @@ export default class OsmRequest {
       ...options
     };
 
-    this._options.endpoint = removeTrailingSlashes(this._options.endpoint);
-
-    if (this._options.basicauth) {
-      this._auth = { basic: this._options.basicauth };
-    } else {
-      this._auth = osmAuth({
-        url: this._options.endpoint,
-        oauth_consumer_key: this._options.oauthConsumerKey,
-        oauth_secret: this._options.oauthSecret,
-        oauth_token: this._options.oauthUserToken,
-        oauth_token_secret: this._options.oauthUserTokenSecret
-      });
-    }
+    this._options.url = removeTrailingSlashes(this._options.url);
+    this._options.apiUrl = removeTrailingSlashes(this._options.apiUrl);
+    this._auth = osmAuth({
+      scope: this._options.scope,
+      client_id: this._options.client_id,
+      url: this._options.url,
+      apiUrl: this._options.apiUrl,
+      redirect_uri: this._options.redirect_uri,
+      access_token: this._options.access_token,
+      auto: this._options.auto,
+      singlepage: this._options.singlepage,
+      loading: this._options.loading,
+      done: this._options.done,
+      locale: this._options.locale
+    });
   }
 
   /**
-   * Return the API endpoint to use for the requests
-   * @return {string} URL of the API endpoint
+   * Return the API URL to use for the requests
+   * @return {string} URL of the API
    */
-  get endpoint() {
-    return this._options.endpoint;
+  get apiUrl() {
+    return this._options.apiUrl;
   }
 
   /**
@@ -91,7 +98,7 @@ export default class OsmRequest {
    */
   fetchNotes(left, bottom, right, top, limit = null, closedDays = null) {
     return fetchNotesRequest(
-      this.endpoint,
+      this.apiUrl,
       left,
       bottom,
       right,
@@ -127,7 +134,7 @@ export default class OsmRequest {
     to = null
   ) {
     return fetchNotesSearchRequest(
-      this.endpoint,
+      this.apiUrl,
       q,
       format,
       limit,
@@ -150,7 +157,7 @@ export default class OsmRequest {
    */
   fetchNote(noteId, format = 'xml') {
     return fetchNoteByIdRequest(
-      this.endpoint,
+      this.apiUrl,
       noteId,
       format,
       this._options.always_authenticated ? { auth: this._auth } : {}
@@ -165,7 +172,7 @@ export default class OsmRequest {
    * @return {Promise}
    */
   createNote(lat, lon, text) {
-    return createNoteRequest(this._auth, this.endpoint, lat, lon, text);
+    return createNoteRequest(this._auth, this.apiUrl, lat, lon, text);
   }
 
   /**
@@ -176,7 +183,7 @@ export default class OsmRequest {
   commentNote(noteId, text) {
     return genericPostNoteRequest(
       this._auth,
-      this.endpoint,
+      this.apiUrl,
       noteId,
       text,
       'comment'
@@ -191,7 +198,7 @@ export default class OsmRequest {
   closeNote(noteId, text) {
     return genericPostNoteRequest(
       this._auth,
-      this.endpoint,
+      this.apiUrl,
       noteId,
       text,
       'close'
@@ -206,7 +213,7 @@ export default class OsmRequest {
   reopenNote(noteId, text) {
     return genericPostNoteRequest(
       this._auth,
-      this.endpoint,
+      this.apiUrl,
       noteId,
       text,
       'reopen'
@@ -223,7 +230,7 @@ export default class OsmRequest {
   createChangeset(createdBy = '', comment = '', tags = {}) {
     return createChangesetRequest(
       this._auth,
-      this.endpoint,
+      this.apiUrl,
       createdBy,
       comment,
       tags
@@ -237,7 +244,7 @@ export default class OsmRequest {
    */
   isChangesetStillOpen(changesetId) {
     return changesetCheckRequest(
-      this.endpoint,
+      this.apiUrl,
       changesetId,
       this._options.always_authenticated ? { auth: this._auth } : {}
     );
@@ -250,7 +257,7 @@ export default class OsmRequest {
    */
   fetchChangeset(changesetId) {
     return changesetGetRequest(
-      this.endpoint,
+      this.apiUrl,
       changesetId,
       this._options.always_authenticated ? { auth: this._auth } : {}
     );
@@ -268,7 +275,7 @@ export default class OsmRequest {
   updateChangesetTags(changesetId, createdBy = '', comment = '', object = {}) {
     return updateChangesetTagsRequest(
       this._auth,
-      this.endpoint,
+      this.apiUrl,
       changesetId,
       createdBy,
       comment,
@@ -283,7 +290,7 @@ export default class OsmRequest {
    * @return {Promise} Empty string if it works
    */
   closeChangeset(changesetId) {
-    return closeChangesetRequest(this._auth, this.endpoint, changesetId);
+    return closeChangesetRequest(this._auth, this.apiUrl, changesetId);
   }
 
   /**
@@ -296,7 +303,7 @@ export default class OsmRequest {
   uploadChangesetOsc(changesetId, osmChangeContent) {
     return uploadChangesetOscRequest(
       this._auth,
-      this.endpoint,
+      this.apiUrl,
       changesetId,
       osmChangeContent
     );
@@ -318,7 +325,7 @@ export default class OsmRequest {
    * @return {Promise}
    */
   fetchChangesets(options) {
-    return fetchChangesetsRequest(this.endpoint, {
+    return fetchChangesetsRequest(this.apiUrl, {
       auth: this._options.always_authenticated ? this._auth : null,
       ...options
     });
@@ -440,13 +447,13 @@ export default class OsmRequest {
   fetchElement(osmId, options) {
     if (options && options.full) {
       return fetchElementRequestFull(
-        this.endpoint,
+        this.apiUrl,
         osmId,
         this._options.always_authenticated ? { auth: this._auth } : {}
       );
     } else {
       return fetchElementRequest(
-        this.endpoint,
+        this.apiUrl,
         osmId,
         this._options.always_authenticated ? { auth: this._auth } : {}
       );
@@ -460,7 +467,7 @@ export default class OsmRequest {
    */
   fetchMultipleElements(osmIds) {
     return multiFetchElementsByTypeRequest(
-      this.endpoint,
+      this.apiUrl,
       osmIds,
       this._options.always_authenticated ? { auth: this._auth } : {}
     );
@@ -473,7 +480,7 @@ export default class OsmRequest {
    */
   fetchRelationsForElement(osmId) {
     return fetchRelationsForElementRequest(
-      this.endpoint,
+      this.apiUrl,
       osmId,
       this._options.always_authenticated ? { auth: this._auth } : {}
     );
@@ -486,7 +493,7 @@ export default class OsmRequest {
    */
   fetchWaysForNode(osmId) {
     return fetchWaysForNodeRequest(
-      this.endpoint,
+      this.apiUrl,
       osmId,
       this._options.always_authenticated ? { auth: this._auth } : {}
     );
@@ -749,7 +756,7 @@ export default class OsmRequest {
    * @return {Promise}
    */
   sendElement(element, changesetId) {
-    return sendElementRequest(this._auth, this.endpoint, element, changesetId);
+    return sendElementRequest(this._auth, this.apiUrl, element, changesetId);
   }
 
   /**
@@ -763,7 +770,7 @@ export default class OsmRequest {
    */
   fetchMapByBbox(left, bottom, right, top, mode = 'json') {
     return fetchMapByBboxRequest(
-      this.endpoint,
+      this.apiUrl,
       left,
       bottom,
       right,
@@ -780,12 +787,7 @@ export default class OsmRequest {
    * @return {Promise} Promise with the new version number due to deletion
    */
   deleteElement(element, changesetId) {
-    return deleteElementRequest(
-      this._auth,
-      this.endpoint,
-      element,
-      changesetId
-    );
+    return deleteElementRequest(this._auth, this.apiUrl, element, changesetId);
   }
 
   /**
@@ -795,7 +797,7 @@ export default class OsmRequest {
    */
   fetchUser(userId) {
     return fetchUserRequest(
-      this.endpoint,
+      this.apiUrl,
       userId,
       this._options.always_authenticated ? { auth: this._auth } : {}
     );
@@ -806,7 +808,7 @@ export default class OsmRequest {
    * @return {Promise} Promise with Well formatted JSON of user preferences
    */
   getUserPreferences() {
-    return getUserPreferencesRequest(this._auth, this.endpoint);
+    return getUserPreferencesRequest(this._auth, this.apiUrl);
   }
 
   /**
@@ -815,7 +817,7 @@ export default class OsmRequest {
    * @return {Promise} Promise
    */
   setUserPreferences(object) {
-    return setUserPreferencesRequest(this._auth, this.endpoint, object);
+    return setUserPreferencesRequest(this._auth, this.apiUrl, object);
   }
 
   /**
@@ -824,7 +826,7 @@ export default class OsmRequest {
    * @return {Promise} Promise with the value for the key
    */
   getUserPreferenceByKey(key) {
-    return getUserPreferenceByKeyRequest(this._auth, this.endpoint, key);
+    return getUserPreferenceByKeyRequest(this._auth, this.apiUrl, key);
   }
 
   /**
@@ -834,7 +836,7 @@ export default class OsmRequest {
    * @return {Promise} Promise
    */
   setUserPreferenceByKey(key, value) {
-    return setUserPreferenceByKeyRequest(this._auth, this.endpoint, key, value);
+    return setUserPreferenceByKeyRequest(this._auth, this.apiUrl, key, value);
   }
 
   /**
@@ -843,6 +845,6 @@ export default class OsmRequest {
    * @return {Promise} Promise
    */
   deleteUserPreference(key) {
-    return deleteUserPreferenceRequest(this._auth, this.endpoint, key);
+    return deleteUserPreferenceRequest(this._auth, this.apiUrl, key);
   }
 }