+++ webpack.config.js
@@ -1,21 +1,8 @@
-const webpack = require('webpack');
 const path = require('path');
 
-const plugins = [
-  new webpack.optimize.UglifyJsPlugin({
-    minimize: true,
-    mangle: true,
-    output: {
-      comments: false
-    },
-    compress: {
-      warnings: false
-    }
-  })
-];
-
 module.exports = {
-  plugins,
+  mode: 'development',
+  plugins: [],
   devtool: 'cheap-module-source-map',
   entry: {
     OsmRequest: path.join(__dirname, 'src/index.js')
@@ -31,7 +18,13 @@ module.exports = {
     modules: [
       'node_modules',
       process.env.NODE_PATH
-    ]
+    ],
+    fallback: {
+      "stream": false,
+      "timers": false,
+      "buffer": false,
+      "string_decoder": false,
+    }
   },
   module: {
     rules: [