+++ CHANGELOG.md
@@ -4,5 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.0.0] - 2024-07-01
+
 ### Added
+
+- Support of OAuth 2.0 via library osm-auth 2.5.0
+
+### Removed
+
+- Support of OAuth 1.0 and basic auth